Initial, dirty proof-of-concept VFD control code.
The RS485 connection uses pins 8 (Tx), 9 (Rx) and 10 (WrEn/~RdEn). This
version talks a Modbus-like protocol used by earlier Huanyang inverter
models, in RTU mode. It supports speed setting, enabling/disabling the
spindle and the spindle-at-speed signal. The Hz-to-RPM conversion is
currently fixed at 30, but it should be made a parameter instead.

Reverse mode doesn't work. I don't know why, might be something inverter
related. Sending a command 0x20 to invert the direction instead seems
to work, but was difficult to get working reliably so I abandoned
the efforts for now.

I haven't tested it with a motor connected to the VFD yet. The Ethernet
side of the implementation is totally hacky and needs to be replaced.