Many fixes and cleanups for Fuling code.
This is based on a mostly-working proof of concept, but hasn't been tested
after the cleanup, so there's some risk of breakage. Also, it still does
not allow a choice of inverter type. It also assumes 800 Hz and 4 poles
for Fuling inverters. There are also issues with excessive delays
and a total lack of error reporting.